2.5.0 Prep
- Added more verbose error/status logging
- Improved Backup/Restore functionality
- Updated D2RHUD.dll for improved reliability
- Fixed Customizations Value-Saving Issue
- Code Cleanup for organization, fixes and simplification

+++ src/D2RLaunch/D2RLaunch/ViewModels/Dialogs/RestoreBackupViewModel.cs
@@ -1,10 +1,7 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.IO;
 using System.Linq;
-using System.Text;
-using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 using System.Windows;
 using System.Windows.Forms;
@@ -12,20 +9,22 @@ using Caliburn.Micro;
 using JetBrains.Annotations;
 using ILog = log4net.ILog;
 using LogManager = log4net.LogManager;
-using WinForms = System.Windows.Forms;
 
 namespace D2RLaunch.ViewModels.Dialogs
 {
     public class RestoreBackupViewModel : Caliburn.Micro.Screen
     {
-        #region members
+        #region ---Static Members---
 
         private ObservableCollection<string> _characters = new ObservableCollection<string>();
         private ILog _logger = LogManager.GetLogger(typeof(RestoreBackupViewModel));
         private string _selectedCharacter;
 
         OpenFileDialog openFileDialog = new OpenFileDialog();
+
         #endregion
+
+        #region ---Window/Loaded Handlers---
 
         public RestoreBackupViewModel(ShellViewModel shellViewModel)
         {
@@ -35,10 +34,11 @@ namespace D2RLaunch.ViewModels.Dialogs
             Execute.OnUIThread(async () => { await GetCharactersToRestore(); });
         }
 
-        #region properties
+        #endregion
+
+        #region ---Properties---
 
         public ShellViewModel ShellViewModel { get; }
-
         public string SelectedCharacter
         {
             get => _selectedCharacter;
@@ -52,7 +52,6 @@ namespace D2RLaunch.ViewModels.Dialogs
                 NotifyOfPropertyChange();
             }
         }
-
         public ObservableCollection<string> Characters
         {
             get => _characters;
@@ -69,44 +68,52 @@ namespace D2RLaunch.ViewModels.Dialogs
 
         #endregion
 
-        private async Task GetCharactersToRestore()
+        #region ---Backup/Restore Functions--
+
+        private async Task GetCharactersToRestore() //Populate dropbox with retrieved backup file list
         {
             if (Directory.Exists(ShellViewModel.BackupFolder))
             {
-                foreach (string character in Directory.GetDirectories(ShellViewModel.BackupFolder).Where(f => !f.Contains("Stash")))
+                foreach (string character in Directory.GetDirectories(ShellViewModel.BackupFolder))
                 {
                     DirectoryInfo directoryInfo = new DirectoryInfo(character);
                     Characters.Add(directoryInfo.Name);
                 }
             }
         }
-
         [UsedImplicitly]
-        public async void OnRestoreBackupSelection()
+        public async void OnRestoreBackupSelection() //Open File Dialog to Backup Folder
         {
-            openFileDialog.ShowDialog();
             openFileDialog.InitialDirectory = Path.Combine(ShellViewModel.BackupFolder, SelectedCharacter);
+            openFileDialog.ShowDialog();
             openFileDialog.Filter = "D2 Save Files (*.d2s)|*.d2s";
         }
-
-            [UsedImplicitly]
-        public async void OnRestoreBackup()
+        [UsedImplicitly]
+        public async void OnRestoreBackup() //Restore chosen stash or character file
         {
             string characterPath = Path.Combine(ShellViewModel.BackupFolder, SelectedCharacter);
             string stashPath = Path.Combine(ShellViewModel.BackupFolder, "Stash");
 
-            if (Directory.Exists(characterPath))
-                File.Copy(openFileDialog.FileName, Path.Combine(ShellViewModel.SaveFilesFilePath, $"{SelectedCharacter}.d2s"), true);
-
-            if (Directory.Exists(stashPath))
+            if (Directory.Exists(characterPath) && (Directory.Exists(stashPath)))
             {
-                File.Copy(Path.Combine(stashPath, "SharedStashSoftCoreV2.d2i"), Path.Combine(ShellViewModel.SaveFilesFilePath, "SharedStashSoftCoreV2.d2i"), true);
-                File.Copy(Path.Combine(stashPath, "SharedStashHardCoreV2.d2i"), Path.Combine(ShellViewModel.SaveFilesFilePath, "SharedStashHardCoreV2.d2i"), true);
+                if (openFileDialog.FileName.Contains(".d2s"))
+                {
+                    File.Copy(openFileDialog.FileName, Path.Combine(ShellViewModel.SaveFilesFilePath, $"{SelectedCharacter}.d2s"), true);
+                    System.Windows.MessageBox.Show($"{SelectedCharacter} Restored!", "Restore Complete", MessageBoxButton.OK, MessageBoxImage.Information);
+                }
+                else
+                {
+                    if (openFileDialog.FileName.Contains("SoftCore"))
+                        File.Copy(openFileDialog.FileName, Path.Combine(ShellViewModel.SaveFilesFilePath, "SharedStashSoftCoreV2.d2i"), true);
+                    else
+                        File.Copy(openFileDialog.FileName, Path.Combine(ShellViewModel.SaveFilesFilePath, "SharedStashHardCoreV2.d2i"), true);
+
+                    System.Windows.MessageBox.Show($"Shared Stash Restored!", "Restore Complete", MessageBoxButton.OK, MessageBoxImage.Information);
+                }
             }
-
-
-            System.Windows.MessageBox.Show($"{SelectedCharacter} Restored!", "Restore Complete", MessageBoxButton.OK, MessageBoxImage.Information);
             await TryCloseAsync();
         }
+
+        #endregion
     }
 }
